// web/src/app/generated/model/rentBookRequest.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RentBookRequest { 
    /**
     * ID of the book to rent
     */
    bookId: string;
    /**
     * ID of the user renting the book
     */
    userId: string;
}