// web/src/app/generated/model/rentBookResponse.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Book } from './book';


export interface RentBookResponse { 
    /**
     * Whether the rental was successful
     */
    success: boolean;
    /**
     * Success or error message
     */
    message: string;
    book?: Book;
    /**
     * Date and time when the book was rented
     */
    rentedDate?: string;
}