// web/src/app/generated/model/returnBookRequest.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReturnBookRequest { 
    /**
     * ID of the book to return
     */
    bookId: string;
    /**
     * ID of the user returning the book
     */
    userId: string;
}