// web/src/app/generated/model/returnBookResponse.ts
/**
 * Modul450 Library API
 * Simple library management system for renting books
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Book } from './book';


export interface ReturnBookResponse { 
    /**
     * Whether the return was successful
     */
    success: boolean;
    /**
     * Success or error message
     */
    message: string;
    book?: Book;
    /**
     * Date and time when the book was returned
     */
    returnedDate?: string;
}